### 2020-01-11 #358 @iecastro
submission: healthdatacsv
Submitting Author Name: Ivan Castro
Submitting Author Github Handle: @iecastro
Repository:   https://github.com/iecastro/healthdatacsv
Version submitted:   0.0.1
Version accepted: TBD

---

```
Package: healthdatacsv
Title: Access data in the healthdata.gov catalog
Version: 0.0.1
    person(given = "Ivan",
           family = "Castro",
           email = "ivedcastr@gmail.com")
Description: This package will make an API call to the
    data.json endpoint of HealthData.gov and read data into R, if available in
    csv format. Some helper functions are available to help users query by
    listed Agency and/or keywords.
    httr,
    vroom
    covr
```

  `healthdatacsv` falls under *data retrieval* because it accesses and ingests data from an online source.  This package focuses on health-related data made available from the [HealthData.gov API](https://healthdata.gov/node/25341).

Target audience would be researchers, instructors, students, and analysts in health-related fields (i.e. Epidemiology, Health Economics, Health Services Research).  The data accessed are collected by agencies of the US Dept. of Health and Human Services, and are often nationally representative. Scientific applications include, among others, the investigation of: disease patterns, trends in health behaviors, or deficits in health systems.

There is [rHealthDataGov](https://github.com/rOpenHealth/rHealthDataGov); however, this package has not been updated and does not function properly with the current [HealthData.gov API](https://healthdata.gov/node/25341).

https://github.com/ropensci/software-review/issues/350
