+++ README.md
@@ -207,29 +207,29 @@ data_raw %>%
 #> $ description    <chr> "<p>This dataset contains policy data for 50 US s…
 #> $ modified       <chr> "2018-09-25", "2018-09-25", "2018-09-25", "2018-0…
 #> $ downloadURL    <chr> "https://data.cdc.gov/api/views/nxst-x9p4/rows.cs…
-#> $ Year           <dbl> 2011, 2011, 2011, 2010, 2015, 2007, 2012, 2010, 2…
-#> $ Quarter        <dbl> 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
-#> $ LocationAbbr   <chr> "TN", "TN", "FL", "DC", "AL", "OK", "MI", "CO", "…
-#> $ LocationDesc   <chr> "Tennessee", "Tennessee", "Florida", "District of…
-#> $ HealthTopic    <chr> "Nutrition", "Obesity", "Obesity", "Nutrition", "…
-#> $ PolicyTopic    <chr> "Media Campaigns", "Food Restrictions", "Incentiv…
+#> $ Year           <dbl> 2015, 2015, 2011, 2009, 2009, 2007, 2010, 2010, 2…
+#> $ Quarter        <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 3, 1…
+#> $ LocationAbbr   <chr> "FL", "NY", "NY", "NY", "MN", "OK", "PA", "NJ", "…
+#> $ LocationDesc   <chr> "Florida", "New York", "New York", "New York", "M…
+#> $ HealthTopic    <chr> "Physical Activity", "Nutrition", "Obesity", "Obe…
+#> $ PolicyTopic    <chr> "Pedestrians/Walking", "School Nutrition", "Appro…
 #> $ DataSource     <chr> "DNPAO", "DNPAO", "DNPAO", "DNPAO", "DNPAO", "DNP…
-#> $ Setting        <chr> "School/After School", "Restaurant/Retail", "Comm…
-#> $ Title          <chr> "Tort Liability and Reform", "Public Funds and Fi…
-#> $ Status         <chr> "Enacted", "Dead", "Dead", "Enacted", "Enacted", …
-#> $ Citation       <chr> "H1151", "H1385", "H455", "R1136", "SB260", "H283…
-#> $ StatusAltValue <dbl> 1, 3, 3, 1, 1, 1, 1, 1, 3, 3, 1, 1, 3, 3, 3, 1, 3…
+#> $ Setting        <chr> "Community", "School/After School", "School/After…
+#> $ Title          <chr> "Florida Shared-Use Nonmotorized Trail Network", …
+#> $ Status         <chr> "Dead", "Introduced", "Dead", "Dead", "Dead", "En…
+#> $ Citation       <chr> "SB1186", "AB544", "A5296", "A7804", "S1484", "H2…
+#> $ StatusAltValue <dbl> 3, 0, 3, 3, 3, 1, 1, 3, 1, 3, 3, 3, 3, 3, 3, 1, 3…
 #> $ DataType       <chr> "Text", "Text", "Text", "Text", "Text", "Text", "…
-#> $ Comments       <chr> "(Abstract - Removes a statutory reference to cer…
-#> $ EnactedDate    <chr> "01/01/2011 12:00:00 AM", NA, NA, "01/01/2010 12:…
-#> $ EffectiveDate  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N…
[…]
