### 2020-01-14 7c23407
add basic use vignette
