+++ vignettes/basic-use.Rmd
@@ -1,0 +1,173 @@
+---
+title: "Basic use of healthdatacsv"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Basic use of healthdatacsv}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  message = FALSE,
+  warning = FALSE
+)
+
+devtools::load_all()
+
+```
+
+
+**healthdatacsv** allows users to query the [healthdata.gov API](https://healthdata.gov/node/25341) catalog and return tidy data frames. This package focuses on the data.json endpoint and will download datasets if available in csv format.  
+
+
+## Querying and reading health data  
+
+```{r}
+library(healthdatacsv)
+```
+
+With **healthdatacsv** you can:  
+
+- Query the API  
+
+  - based on keyword(s)  
+
+```{r example}
+fetch_catalog(keyword = "alcohol|drugs")
+
+```
+
+
+  - based on agency name  
+  
+```{r}
+fetch_catalog("National Institutes of Health (NIH)")
+
+```
+
+- You can also query the API and read the data into R:  
+
+```{r}
+fetch_catalog("Centers for Disease Control and Prevention",
+                               keyword = "built environment") %>% 
+  fetch_csv()
+
+```
+
+
+## Basic workflow
+
+**healthdatacsv** is pipe friendly, so it's easy to integrate into a tidy workflow. For example, say you're interested in searching CDC datasets related to the built environment.  
+
+Let's fetch the catalog of available data products:
+
+```{r}
+cdc_built_env <- fetch_catalog("Centers for Disease Control and Prevention",
+                               keyword = "built environment")
+
+cdc_built_env
+
+```
+
+In this case, there is only one product available.  To learn more about the dataset, you can simply `pull` the description:
+
+```{r}
+cdc_built_env %>%
+  dplyr::pull(description)
+
+```
+
+This dataset relates to state legislation on nutrition, physical activity, and obesity during 2001-2017.  Data only includes enacted legislation.
+
+To download the data, we pass the catalog df to the `fetch_csv` function.  Data are nested in a list-column, and since there is only one dataset to download, we can `unnest` in the same pipe.  
+
+```{r}
+data_raw <- cdc_built_env %>%
+  fetch_csv() %>% #> 
+  tidyr::unnest(data_tbl)
+
+```
+
+The data are now in R, and ready to be utilized  
+
+```{r}
+data_raw %>%
+  dplyr::glimpse()
+
+```
+
+### Fetching multiple datasets  
+
+Since data are nested into list-columns, is easy to fetch multiple datasets with one call.    
+
+For example, let's query the catalog for data related to reproductive health:    
+
+```{r} 
+library(dplyr) # for table manipulation verbs
+
+# PRAMS data
+# CDC surveillance for reproductive health
+fetch_catalog(keyword = "reproductive health")
+
+```
+
+Here we can use dplyr verbs to help us manipulate the table and fetch data newer than 2009.  
+
+```{r}
+# query, filter, and fetch
+prams <- fetch_catalog(keyword = "reproductive health") %>% 
+  mutate(year = readr::parse_number(product)) %>% 
+  filter(year > 2009) %>% 
+  arrange(year) %>% 
+  fetch_csv()
+
+```
+
+The data are now available for you to use.  Thanks to the [vroom](https://vroom.r-lib.org/index.html) package, the data are indexed, and memory allocation is saved until you actually access (unnest) the data.  
+
+```{r}
+prams %>% 
+  select(product, data_tbl)
+```
+
+
+### Some helper functions
+
[…]
