+++ DESCRIPTION
@@ -27,7 +27,10 @@ Imports:
 Suggests: 
     readr,
     testthat (>= 2.1.0),
-    covr
+    covr,
+    knitr,
+    rmarkdown
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.1
+VignetteBuilder: knitr
