### 2020-01-14 fe520dc
remove devtools dep from vignette
