### 2020-01-14 #358 @iecastro
Hi all: did a small update to include a COC, and a basic use vignette.
