### 2020-01-24 #358 @lmullen
### Editor checks:

---

#### Editor comments

Thanks for your submission. I have requests out to reviewers, and will let you know the reviewers and the review date when enough have agreed.

In the meantime, could you please address these issues?

- [ ] The repository uses codecov but the automated code coverage through Travis CI functionality is not currently working.
- [ ] A few mis-spellings: `healdata`
- [ ] Can you add CI for Windows with Appveyor? Try the `usethis` package.

---

Reviewers:
Due date:
