### 2020-01-24 #358 @lmullen
One other thing: while the package is undergoing peer review, could you please add the peer review badge? You can do this by running `rodeo::use_review_badge(358)`.
