+++ README.md
@@ -11,6 +11,7 @@ maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www
 status](https://travis-ci.org/iecastro/healthdatacsv.svg?branch=master)](https://travis-ci.org/iecastro/healthdatacsv)
 [![Codecov test
 coverage](https://codecov.io/gh/iecastro/healthdatacsv/branch/master/graph/badge.svg)](https://codecov.io/gh/iecastro/healthdatacsv?branch=master)
+[![peer-review](https://badges.ropensci.org/358_status.svg)](https://github.com/ropensci/software-review/issues/358)
 <!-- badges: end -->
 
 **healthdatacsv** allows users to query the [healthdata.gov
@@ -45,11 +46,11 @@ fetch_catalog(keyword = "alcohol|drugs")
 #> # A tibble: 137 x 6
 #>    publisher.name   product   description   modified distribution csv_avail
 #>    <chr>            <chr>     <chr>         <chr>    <list>       <lgl>    
-#>  1 Centers for Dis… Alzheime… "<p>2011-201… 2019-05… <df[,6] [4 … TRUE     
-#>  2 U.S. Food and D… All FDA … "<p>Contains… 2010-10… <df[,6] [1 … FALSE    
-#>  3 U.S. Food and D… Adverse … "<p>The Adve… 2004-01… <df[,6] [1 … FALSE    
-#>  4 U.S. Food and D… Drugs to… "<p>Companie… 2013-09… <df[,6] [1 … FALSE    
-#>  5 Centers for Dis… U.S. Chr… "<p>CDC's Di… 2018-07… <df[,6] [4 … TRUE     
+#>  1 Centers for Dis… U.S. Chr… "<p>CDC's Di… 2018-07… <df[,6] [4 … TRUE     
+#>  2 Centers for Dis… Alzheime… "<p>2011-201… 2019-05… <df[,6] [4 … TRUE     
+#>  3 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
+#>  4 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
+#>  5 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
 #>  6 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
 #>  7 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
 #>  8 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
@@ -85,8 +86,8 @@ fetch_catalog(keyword = "alcohol")
 #> # A tibble: 123 x 6
 #>    publisher.name   product   description   modified distribution csv_avail
 #>    <chr>            <chr>     <chr>         <chr>    <list>       <lgl>    
-#>  1 Centers for Dis… Alzheime… "<p>2011-201… 2019-05… <df[,6] [4 … TRUE     
-#>  2 Centers for Dis… U.S. Chr… "<p>CDC's Di… 2018-07… <df[,6] [4 … TRUE     
+#>  1 Centers for Dis… U.S. Chr… "<p>CDC's Di… 2018-07… <df[,6] [4 … TRUE     
+#>  2 Centers for Dis… Alzheime… "<p>2011-201… 2019-05… <df[,6] [4 … TRUE     
 #>  3 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
 #>  4 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
[…]
