+++ .travis.yml
@@ -5,4 +5,4 @@ cache: packages
 sudo: required
 
 after_success:
-  - Rscript -e 'covr::codecov()
+  - Rscript -e 'covr::codecov()'
