### 2020-02-08 5a1e289
added appveyor ci badge, and yml
