+++ README.md
@@ -12,6 +12,8 @@ status](https://travis-ci.org/iecastro/healthdatacsv.svg?branch=master)](https:/
 [![Codecov test
 coverage](https://codecov.io/gh/iecastro/healthdatacsv/branch/master/graph/badge.svg)](https://codecov.io/gh/iecastro/healthdatacsv?branch=master)
 [![peer-review](https://badges.ropensci.org/358_status.svg)](https://github.com/ropensci/software-review/issues/358)
+[![AppVeyor build
+status](https://ci.appveyor.com/api/projects/status/github/iecastro/healthdatacsv?branch=master&svg=true)](https://ci.appveyor.com/project/iecastro/healthdatacsv)
 <!-- badges: end -->
 
 **healthdatacsv** allows users to query the [healthdata.gov
@@ -46,16 +48,16 @@ fetch_catalog(keyword = "alcohol|drugs")
 #> # A tibble: 137 x 6
 #>    publisher.name   product   description   modified distribution csv_avail
 #>    <chr>            <chr>     <chr>         <chr>    <list>       <lgl>    
-#>  1 Centers for Dis… U.S. Chr… "<p>CDC's Di… 2018-07… <df[,6] [4 … TRUE     
-#>  2 Centers for Dis… Alzheime… "<p>2011-201… 2019-05… <df[,6] [4 … TRUE     
-#>  3 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
-#>  4 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
-#>  5 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
-#>  6 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
-#>  7 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
-#>  8 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
-#>  9 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
-#> 10 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,6] [4 … TRUE     
+#>  1 Centers for Dis… Alzheime… "<p>2011-201… 2020-02… <df[,5] [4 … TRUE     
+#>  2 Centers for Dis… U.S. Chr… "<p>CDC's Di… 2020-01… <df[,5] [4 … TRUE     
+#>  3 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
+#>  4 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
+#>  5 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
+#>  6 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
+#>  7 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
+#>  8 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
+#>  9 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
+#> 10 Centers for Dis… CDC PRAM… "<ol start=\… 2018-07… <df[,5] [4 … TRUE     
[…]
