+++ .travis.yml
@@ -1,6 +1,11 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 language: R
+r:
+  - oldrel
+  - release
+  - devel
+
 cache: packages
 sudo: required
 
