+++ appveyor.yml
@@ -22,6 +22,15 @@ environment:
 
 # Adapt as necessary starting from here
 
+matrix:
+  - R_VERSION: devel
+    R_ARCH: x64
+
+  - R_VERSION: release
+    R_ARCH: x64
+
+  - R_VERSION: patched
+
 build_script:
   - travis-tool.sh install_deps
 
