### 2020-02-08 ddb748c
should fix appveyor parsing error
