### 2020-02-08 #358 @iecastro
> ### Editor checks:
> * [x]  **Fit**: The package meets criteria for [fit](policies.md#fit) and [overlap](policies.md#fit)
> * [x]  **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service.
> * [x]  **License:** The package has a CRAN or OSI accepted license
> * [x]  **Repository:** The repository link resolves correctly
>
> #### Editor comments
> Thanks for your submission. I have requests out to reviewers, and will let you know the reviewers and the review date when enough have agreed.
>
> In the meantime, could you please address these issues?
>
> * [x]  The repository uses codecov but the automated code coverage through Travis CI functionality is not currently working.
> * [ ]  A few mis-spellings: `healdata`
    - I couldn't find these, but will keep in mind
> * [x]  Can you add CI for Windows with Appveyor? Try the `usethis` package.
>
> Reviewers:
> Due date:
