### 2020-02-26 #358 @lmullen
Thanks to our reviewers for agreeing to take on the peer review of this package.

Reviewer: @czeildi
Reviewer: @cksun-usc
Due date: 2020-03-19

You can find details about how to do the review in the [reviewers' guide](https://devguide.ropensci.org/reviewerguide.html).
