### 2020-03-07 #358 @czeildi
* in `fetch_csv()`, prefer `dplyr::slice()` to `slice()` so it runs without `library(dplyr)`

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

In general, I found the package easy to learn and use, and the code is readable and simple. Below are my comments in three categories. Most of them are non-blocking.

- [ ] All functions calls emit the message: `"No encoding supplied: defaulting to UTF-8."` Consider muting this with `encoding = "UTF-8"` in `httr::content(response, "text")` in `healthdata_api()`.
- [ ] `fetch_csv()` throws a bit confusing error when called with a catalog where `csv_available` is `FALSE`. I would add and explicit test and error message for this, and also highlight in the function documentation. (Also consider the case of multiple catalogs, where some of the have `csv_available`, while others do not.)
- [ ] The `data_viewer` argument in `get_keywords()` seems arbitrary for me: why here and not in other functions? It makes the code more complicated for little gain for the user (they have to provide an argument instead of for example `%T>%` into `View()`). I would suggest to remove this argument.
- [ ] 3 different verbs are used for similar things: getting results from the API (get, fetch, list). Is there a logic to make this easier to remember? I would use the same verb for keywords and agencies as they are both helpers so that the user knows what arguments they can call `fetch_catalog()` with.
- [ ] consider renaming `fetch_csv()`: it does not return csv files so it was a bit confusing for me
- [ ] in `list_agencies` the `namecheck` argument is not really intuitive to me, possibly use `name_match` instead or sg suggesting regex matching?
- [ ] Besides `fetch_csv()` all functions (with any arguments) call `healthdata_api("data.json")` which takes a significant time. This could be better for the user but also makes tests and vignette compilation quite slow. Consider caching the result of the api call or reorganize functions in a way that the user can easily re-use the api result. I believe the [memoise](https://cran.r-project.org/package=memoise) package can be used for this but I do not have experience with it, it might not worth it in this case.

#### code style

- [ ] Use consistent indentation and spacing (You can do this automatically with `styler::style_pkg()`.)
- [ ] consider renaming `publisher.name` to `publisher_name` to follow snake case in results as well
- [ ] In `healthdata_api` if there is an error, you cannot necessarily parse so I believe the error message should not rely on the parsed response in all cases (see `healthdata_api("data.jsn")` as a forced example)
- [ ] add a new test for the case of error from the API
- [ ] There are repeated parts in `fetch_catalog()`, `list_agencies()` and `get_keywords()` which could be simplified and thus easier to understand and maintain. E.g. in `get_keywords()`:
```r
keywords <- jsonlite::flatten(parsed_dataset) %>%
  as_tibble() %>%
  select(
    .data$publisher.name,
    .data$keyword
  ) %>%
  tidyr::unnest(cols = c(.data$keyword)) %>%
  distinct()
if (!is.null(agency)) {
  keywords <- keywords %>%
    filter(.data$publisher.name == agency)
}
```
- [ ] it is good practice to give tests unique descriptions so that if there is a failure, it is easier to find based on description. (In this case there are only a few tests so it is ok). Nonetheless, consider a  unique description for all test cases (you can achieve this by changing the descriptions or grouping similar expectations together in one test case)
- [ ] test files: it is not necessary to include `library()` call at the top of the test files (I am not sure whether it is problematic, but it is not standard afaik)
- [ ] there is no need for explicit `return()` calls, consider removing them
- [ ] in example in `fetch_csv()`, prefer `dplyr::slice()` to `slice()` so it runs without `library(dplyr)`

- [ ] add `Roxygen: list(markdown = TRUE)` to DESCRIPTION so that markdown formatting works in roxygen2 documentation
- [ ] use `@noRd` instead of `@keyword internal` for documenting internal functions. [Reference 1](https://devguide.ropensci.org/building.html#documentation), [Reference 2](https://community.rstudio.com/t/keywords-internal-vs-nord/35119/2)
- [ ] consider adding codemetar: https://devguide.ropensci.org/building.html#package-name-and-metadata
- [ ] The readme is a bit long to me compared to the number of functionalities. Consider using `head` on results. (Note that I am also completely fine with the current readme)
- [ ] Add a repostatus.org badge to readme as per guideline in https://devguide.ropensci.org/building.html#readme
- [ ] Fix link to `tibble` in function documentation `@return` part: use ``[`tibble`][tibble::tibble()]`` or similar. (I checked in `pkgdown` documentation, but did not find a way to link to tibble-package documentation without hardlinging url-s)
- [ ] Separate titles and description in package documentation documentation (insert blank line or add separate title) so that the title is short, especially in `fetch_catalog()`
- [ ] use consistent code style (indentation) in readme, vignette as well, e.g.
```r
fetch_catalog("Centers for Disease Control and Prevention",
                               keyword = "built environment")
```
- [ ] There is a sentence in the readme: "healthdatacsv is pipe friendly, so it’s easy to integrate into a tidy workflow." It is a bit out of context for me: this claim is not really supported, the single command pipeline with `dplyr::pull` is not a strong argument and all functions in the package have different input an output. (This is not to say that the package is pipe-unfriendly!). The `fetch_catalog(...) %>% fetch_csv()` pipe supports this sentence more.
- [ ] indicate all internal functions with `function_name()` in readme and vignette for nice linking in `pkgdown` website
- [ ] in function documentations: note NULL, TRUE as code (by enclosing with \`\`), healthdata.gov as link, link to other functions by enclosing in \`\`
