### 2020-03-09 #358 @iecastro
Thanks @czeildi!  I appreciate the feedback and will start addressing individual items soon.
