### 2020-03-10 b2c14f4
started on reviewer 1 usability and code style  comments
