+++ R/healthdatacsv-package.R
@@ -6,7 +6,7 @@
 #' @importFrom dplyr as_tibble select mutate filter distinct
 #' @importFrom httr user_agent modify_url GET http_error status_code
 #' @importFrom jsonlite fromJSON flatten
-#' @importFrom stringr str_detect
+#' @importFrom stringr str_detect regex
 #' @importFrom purrr map map_lgl
 #' @importFrom tidyr unnest
 #' @importFrom tibble tibble
