+++ NAMESPACE
@@ -21,6 +21,7 @@ importFrom(magrittr,"%>%")
 importFrom(purrr,map)
 importFrom(purrr,map_lgl)
 importFrom(rlang,.data)
+importFrom(stringr,regex)
 importFrom(stringr,str_detect)
 importFrom(tibble,tibble)
 importFrom(tidyr,unnest)
