### 2020-03-10 385ee70
start on rvwr-1 documentation, fixed function and test, update vignette
