+++ R/helpers.R
@@ -4,7 +4,7 @@
 #'        catalog for string pattern.
 #'        Defaults to NULL and pulls all agencies cataloged in API
 #'
-#' @return a [tibble][tibble::tibble-package] of names
+#' @return a [`tibble`][tibble::tibble()] of names
 #' @examples
 #' \dontrun{
 #' list_agencies()
@@ -48,8 +48,7 @@ list_agencies <- function(agency = NULL) {
 #' @param agency enter full agency name in title case (results from
 #'        list_agencies()) to pull keywords tagged in the listed
 #'        agency's products. Defaults to NULL and pulls all keywords cataloged
-#' @param data_viewer if TRUE, keywords are loaded to the data viewer
-#' @return a [tibble][tibble::tibble-package] with publisher (agency)
+#' @return a [`tibble`][tibble::tibble()] with publisher (agency)
 #'         name(s) and respective keywords
 #' @examples
 #' \dontrun{
