+++ man/healthdata_api.Rd
@@ -1,12 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/healthdata-api.R
-\name{healthdata_api}
-\alias{healthdata_api}
-\title{healdata.gov API call for data.json endpoint}
-\usage{
-healthdata_api(path)
-}
-\description{
-healdata.gov API call for data.json endpoint
-}
-\keyword{internal}
