+++ DESCRIPTION
@@ -32,5 +32,6 @@ Suggests:
     rmarkdown
 Encoding: UTF-8
 LazyData: true
+Roxygen: list(markdown = TRUE)
 RoxygenNote: 6.1.1
 VignetteBuilder: knitr

[…]
