+++ README.Rmd
@@ -62,7 +62,7 @@ fetch_catalog(keyword = "alcohol|drugs")
 ```{r}
 fetch_catalog("Centers for Disease Control and Prevention",
                                keyword = "built environment") %>% 
-  fetch_csv()
+  fetch_data()
 
 ```
 
@@ -77,11 +77,11 @@ fetch_catalog(keyword = "alcohol")
 # fetch data
 fetch_catalog(keyword = "alcohol") %>% 
   dplyr::slice(1) %>% # dplyr 
-  fetch_csv()
+  fetch_data()
 
 ```
 
-`fetch_csv()` wraps the [vroom](https://vroom.r-lib.org/) function, which helps quickly read relatively large delimited files:  
+`fetch_data()` wraps the [vroom](https://vroom.r-lib.org/) function, which helps quickly read relatively large delimited files:  
 
 ```{r}
 # PRAMS data
@@ -91,7 +91,7 @@ prams <- fetch_catalog(keyword = "reproductive health") %>%
   mutate(year = readr::parse_number(product)) %>% 
   filter(year >= 2010) %>% 
   arrange(year) %>% 
-  fetch_csv()
+  fetch_data()
 
 prams %>% 
   select(product, data_tbl)
@@ -121,12 +121,12 @@ cdc_built_env %>%
 
 This dataset relates to state legislation on nutrition, physical activity, and obesity during 2001-2017.  Data only includes enacted legislation.
 
-To download the data, we pass the catalog object to the `fetch_csv` function.  Since there is only one dataset to download, we can `unnest` in the same pipe.  *If the catalog has more than one product that you'd like to keep, it is recommended to unnest each product separately.* If the catalog consists of several time points of the same dataset, you could unnest in one pipe, given all column names are the same. 
+To download the data, we pass the catalog object to the `fetch_data` function.  Since there is only one dataset to download, we can `unnest` in the same pipe.  *If the catalog has more than one product that you'd like to keep, it is recommended to unnest each product separately.* If the catalog consists of several time points of the same dataset, you could unnest in one pipe, given all column names are the same. 
 
 
 ```{r}
 data_raw <- cdc_built_env %>%
-  fetch_csv() %>% #> 
+  fetch_data() %>% #> 
   tidyr::unnest(data_tbl)
 
 data_raw
