+++ tests/testthat/test-fetch_catalog.R
@@ -1,4 +1,3 @@
-library(healthdatacsv)
 
 test_that("output is tibble", {
   expect_is(fetch_catalog(), "tbl_df")
