+++ tests/testthat/test-list_agencies.R
@@ -1,4 +1,3 @@
-library(healthdatacsv)
 
 test_that("output is tibble", {
   expect_is(list_agencies(), "tbl_df")
