+++ .travis.yml
@@ -1,10 +1,21 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
-language: R
+language: r
+
 r:
   - oldrel
   - release
   - devel
+
+os:
+  - linux
+  - osx
+
+matrix:
+  fast_finish: true
+  allow_failures:
+    - r: devel
+      os: osx
 
 cache: packages
 sudo: required
