+++ NAMESPACE
@@ -10,7 +10,8 @@ importFrom(dplyr,distinct)
 importFrom(dplyr,filter)
 importFrom(dplyr,mutate)
 importFrom(dplyr,select)
-importFrom(httr,GET)
+importFrom(httpcache,GET)
+importFrom(httpcache,startLog)
 importFrom(httr,http_error)
 importFrom(httr,modify_url)
 importFrom(httr,status_code)
