### 2020-07-15 e007f6e
move test coverage to appveyor
