+++ appveyor.yml
@@ -66,3 +66,6 @@ artifacts:
 
   - path: '\*_*.zip'
     name: Bits
+
+on_success:
+  - Rscript -e "covr::codecov()"
