+++ tests/testthat/test-get_agencies.R
@@ -1,10 +1,10 @@
 
 test_that("output is tibble", {
-  expect_is(list_agencies(), "tbl_df")
+  expect_is(get_agencies(), "tbl_df")
 })
 
 
 test_that("output is tibble", {
-  expect_is(list_agencies("CDC"),
+  expect_is(get_agencies("CDC"),
             "tbl_df")
 })
