+++ man/get_agencies.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/helpers.R
-\name{list_agencies}
-\alias{list_agencies}
+\name{get_agencies}
+\alias{get_agencies}
 \title{Extract names of agencies cataloged in the healthdata.gov API}
 \usage{
-list_agencies(agency = NULL)
+get_agencies(agency = NULL)
 }
 \arguments{
 \item{agency}{enter partial name or initials of agency to search
@@ -19,10 +19,10 @@ Extract names of agencies cataloged in the healthdata.gov API
 }
 \examples{
 \dontrun{
-list_agencies()
+get_agencies()
 
-list_agencies("institute")
+get_agencies("institute")
 
-list_agencies("substance abuse")
+get_agencies("substance abuse")
 }
 }
