+++ man/fetch_catalog.Rd
@@ -9,7 +9,7 @@ fetch_catalog(agency = NULL, keyword = NULL)
 \arguments{
 \item{agency}{name of agency to pull in API call.
 This requires the full agency name in proper case.
-Function list_agencies() can be used for a list of
+Function get_agencies() can be used for a list of
 available agencies. Defaults to NULL.}
 
 \item{keyword}{keyword(s) to pull in API call.
