+++ NAMESPACE
@@ -3,8 +3,8 @@
 export("%>%")
 export(fetch_catalog)
 export(fetch_data)
+export(get_agencies)
 export(get_keywords)
-export(list_agencies)
 importFrom(dplyr,as_tibble)
 importFrom(dplyr,distinct)
 importFrom(dplyr,filter)
