+++ README.md
@@ -45,20 +45,22 @@ Basic examples which show you how to use **healthdatacsv**:
 
 ``` r
 fetch_catalog(keyword = "alcohol|drugs")
+#> 2020-07-15T17:18:37.405 HTTP GET http://www.healthdata.gov/data.json 200 477047 0.598 0.216 0.316 0.437 0.695 0.921
+#> 2020-07-15T17:18:37.406 CACHE SET http://www.healthdata.gov/data.json
 #> # A tibble: 134 x 6
-#>    publisher    product    description      modified distribution csv_avail
-#>    <chr>        <chr>      <chr>            <chr>    <list>       <lgl>    
-#>  1 Centers for… Alzheimer… "<p>2011-2017. … 2020-02… <df[,5] [4 … TRUE     
-#>  2 Centers for… U.S. Chro… "<p>CDC's Divis… 2020-01… <df[,5] [4 … TRUE     
-#>  3 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#>  4 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#>  5 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#>  6 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#>  7 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#>  8 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#>  9 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#> 10 Centers for… CDC PRAMS… "<ol start=\"20… 2018-07… <df[,5] [4 … TRUE     
-#> # … with 124 more rows
+#>    publisher     product     description         modified distribution csv_avail
+#>    <chr>         <chr>       <chr>               <chr>    <list>       <lgl>    
+#>  1 Centers for ~ Alzheimer'~ "<p>2011-2017. Thi~ 2020-02~ <df[,5] [4 ~ TRUE     
+#>  2 Centers for ~ U.S. Chron~ "<p>CDC's Division~ 2020-01~ <df[,5] [4 ~ TRUE     
+#>  3 Centers for ~ CDC PRAMSt~ "<ol start=\"2011\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#>  4 Centers for ~ CDC PRAMSt~ "<ol start=\"2000\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#>  5 Centers for ~ CDC PRAMSt~ "<ol start=\"2001\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#>  6 Centers for ~ CDC PRAMSt~ "<ol start=\"2002\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#>  7 Centers for ~ CDC PRAMSt~ "<ol start=\"2003\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#>  8 Centers for ~ CDC PRAMSt~ "<ol start=\"2004\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#>  9 Centers for ~ CDC PRAMSt~ "<ol start=\"2005\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#> 10 Centers for ~ CDC PRAMSt~ "<ol start=\"2006\~ 2018-07~ <df[,5] [4 ~ TRUE     
+#> # ... with 124 more rows
 ```
 
   - Querying API and downloading data for single product:
[…]
