+++ appveyor.yml
@@ -31,8 +31,8 @@ environment:
     R_ARCH: x64
     CURL_SSL_BACKEND: openssl
 
-  - R_VERSION: patched
-    CURL_SSL_BACKEND: openssl
+  #- R_VERSION: patched
+  #  CURL_SSL_BACKEND: openssl
 
 
 matrix:
