### 2020-07-15 7f98695
update doc per review
