+++ R/healthdata-api.R
@@ -5,12 +5,12 @@
 #'
 #' @param agency name of agency to pull in API call.
 #'        This requires the full agency name in proper case.
-#'        Function get_agencies() can be used for a list of
+#'        Function `get_agencies()` can be used for a list of
 #'        available agencies. Defaults to NULL.
 #' @param keyword keyword(s) to pull in API call.
 #'        This argument can be supplied on its own or with an agency argument.
 #'        The keyword argument uses regular expression to detect presence of
-#'        pattern in a string. Function get_keywords() can be used for a list
+#'        pattern in a string. Function `get_keywords()` can be used for a list
 #'        of exact keywords available in API. Defaults to NULL.
 #' @return a [`tibble`][tibble::tibble()] with descriptive metadata of
 #'         the available catalog including a nested list-column of the
@@ -77,11 +77,12 @@ fetch_catalog <- function(agency = NULL,
 
 #' Fetch data from the healthdata.gov API
 #'
-#' This function will download data healthdata.gov,
+#' This function will download data from
+#' [healthdata.gov](https://healthdata.gov/),
 #' if available in csv format
 #'
 #' @param catalog this argument requires a catalog (tibble) created
-#'        with fetch_catalog()
+#'        with `fetch_catalog()`
 #' @return a [`tibble`][tibble::tibble()] with descriptive metadata
 #'         of a data product and a list-column of the available dataset.
 #' @examples
