+++ R/helpers.R
@@ -2,7 +2,7 @@
 #'
 #' @param agency enter partial name or initials of agency to search
 #'        catalog for string pattern.
-#'        Defaults to NULL and pulls all agencies cataloged in API
+#'        Defaults to `NULL` and pulls all agencies cataloged in API
 #'
 #' @return a [`tibble`][tibble::tibble()] of names
 #' @examples
@@ -48,8 +48,8 @@ get_agencies <- function(agency = NULL) {
 #' Extract keywords listed in the healthdata.gov API
 #'
 #' @param agency enter full agency name in title case (results from
-#'        get_agencies()) to pull keywords tagged in the listed
-#'        agency's products. Defaults to NULL and pulls all keywords cataloged
+#'        `get_agencies()`) to pull keywords tagged in the listed
+#'        agency's products. Defaults to `NULL` and pulls all keywords cataloged
 #' @return a [`tibble`][tibble::tibble()] with publisher (agency)
 #'         name(s) and respective keywords
 #' @examples
