+++ man/fetch_catalog.Rd
@@ -9,13 +9,13 @@ fetch_catalog(agency = NULL, keyword = NULL)
 \arguments{
 \item{agency}{name of agency to pull in API call.
 This requires the full agency name in proper case.
-Function get_agencies() can be used for a list of
+Function \code{get_agencies()} can be used for a list of
 available agencies. Defaults to NULL.}
 
 \item{keyword}{keyword(s) to pull in API call.
 This argument can be supplied on its own or with an agency argument.
 The keyword argument uses regular expression to detect presence of
-pattern in a string. Function get_keywords() can be used for a list
+pattern in a string. Function \code{get_keywords()} can be used for a list
 of exact keywords available in API. Defaults to NULL.}
 }
 \value{
