### 2020-07-15 #358 @iecastro
Hi @czeildi - hope all is well!  I've addressed your comments below.  Thank you for your time.

> ### Review Comments
> In general, I found the package easy to learn and use, and the code is readable and simple. Below are my comments in three categories. Most of them are non-blocking.
>
> #### usability
> * [x]  All functions calls emit the message: `"No encoding supplied: defaulting to UTF-8."` Consider muting this with `encoding = "UTF-8"` in `httr::content(response, "text")` in `healthdata_api()`.
> * [x]  `fetch_csv()` throws a bit confusing error when called with a catalog where `csv_available` is `FALSE`. I would add and explicit test and error message for this, and also highlight in the function documentation. (Also consider the case of multiple catalogs, where some of the have `csv_available`, while others do not.)
> * [x]  The `data_viewer` argument in `get_keywords()` seems arbitrary for me: why here and not in other functions? It makes the code more complicated for little gain for the user (they have to provide an argument instead of for example `%T>%` into `View()`). I would suggest to remove this argument.
> * [x]  3 different verbs are used for similar things: getting results from the API (get, fetch, list). Is there a logic to make this easier to remember? I would use the same verb for keywords and agencies as they are both helpers so that the user knows what arguments they can call `fetch_catalog()` with.
  - keywords and agencies now have the same verb *get_*
> * [x]  consider renaming `fetch_csv()`: it does not return csv files so it was a bit confusing for me
  - renamed `fetch_data()`
> * [x]  in `list_agencies` the `namecheck` argument is not really intuitive to me, possibly use `name_match` instead or sg suggesting regex matching?
  - argument renamed `agency`
> * [x]  Besides `fetch_csv()` all functions (with any arguments) call `healthdata_api("data.json")` which takes a significant time. This could be better for the user but also makes tests and vignette compilation quite slow. Consider caching the result of the api call or reorganize functions in a way that the user can easily re-use the api result. I believe the [memoise](https://cran.r-project.org/package=memoise) package can be used for this but I do not have experience with it, it might not worth it in this case.
  - results are now cached using [httpcache](https://github.com/nealrichardson/httpcache)
>
> #### code style
> * [x]  Use consistent indentation and spacing (You can do this automatically with `styler::style_pkg()`.)
> * [x]  consider renaming `publisher.name` to `publisher_name` to follow snake case in results as well
  - renamed to `publisher`
> * [ ]  In `healthdata_api` if there is an error, you cannot necessarily parse so I believe the error message should not rely on the parsed response in all cases (see `healthdata_api("data.jsn")` as a forced example)
  - `healthdata_api` is an internal function, so the user does not have a chance of misspelling the endpoint url.  In theory, there will always be an API call made, and any errors would be parsed.  But also, not much experience on this, on my part.
> * [ ]  add a new test for the case of error from the API
> * [x]  There are repeated parts in `fetch_catalog()`, `list_agencies()` and `get_keywords()` which could be simplified and thus easier to understand and maintain. E.g. in `get_keywords()`:
  - thank you!
>
> ```r
> keywords <- jsonlite::flatten(parsed_dataset) %>%
>   as_tibble() %>%
>   select(
>     .data$publisher.name,
>     .data$keyword
>   ) %>%
>   tidyr::unnest(cols = c(.data$keyword)) %>%
>   distinct()
> if (!is.null(agency)) {
>   keywords <- keywords %>%
>     filter(.data$publisher.name == agency)
> }
> ```
>
> * [x]  it is good practice to give tests unique descriptions so that if there is a failure, it is easier to find based on description. (In this case there are only a few tests so it is ok). Nonetheless, consider a  unique description for all test cases (you can achieve this by changing the descriptions or grouping similar expectations together in one test case)
> * [x]  test files: it is not necessary to include `library()` call at the top of the test files (I am not sure whether it is problematic, but it is not standard afaik)
> * [x]  there is no need for explicit `return()` calls, consider removing them
> * [x]  in example in `fetch_csv()`, prefer `dplyr::slice()` to `slice()` so it runs without `library(dplyr)`
>
> #### documentation
> * [x]  add `Roxygen: list(markdown = TRUE)` to DESCRIPTION so that markdown formatting works in roxygen2 documentation
> * [x]  use `@noRd` instead of `@keyword internal` for documenting internal functions. [Reference 1](https://devguide.ropensci.org/building.html#documentation), [Reference 2](https://community.rstudio.com/t/keywords-internal-vs-nord/35119/2)
> * [ ]  consider adding codemetar: https://devguide.ropensci.org/building.html#package-name-and-metadata
  - Pending. Didn't realize I hadn't looked into this.
> * [x]  The readme is a bit long to me compared to the number of functionalities. Consider using `head` on results. (Note that I am also completely fine with the current readme)
  -  Did not use `head()` but I did revise the readme several times - cutting down the lenght (I think - it's been a while)
> * [x]  Add a repostatus.org badge to readme as per guideline in https://devguide.ropensci.org/building.html#readme
> * [x]  Fix link to `tibble` in function documentation `@return` part: use `` [`tibble`][tibble::tibble()] `` or similar. (I checked in `pkgdown` documentation, but did not find a way to link to tibble-package documentation without hardlinging url-s)
> * [x]  Separate titles and description in package documentation documentation (insert blank line or add separate title) so that the title is short, especially in `fetch_catalog()`
> * [x]  use consistent code style (indentation) in readme, vignette as well, e.g.
>
> ```r
> fetch_catalog("Centers for Disease Control and Prevention",
>                                keyword = "built environment")
> ```
>
> * [x]  There is a sentence in the readme: "healthdatacsv is pipe friendly, so it’s easy to integrate into a tidy workflow." It is a bit out of context for me: this claim is not really supported, the single command pipeline with `dplyr::pull` is not a strong argument and all functions in the package have different input an output. (This is not to say that the package is pipe-unfriendly!). The `fetch_catalog(...) %>% fetch_csv()` pipe supports this sentence more.
> * [x]  indicate all internal functions with `function_name()` in readme and vignette for nice linking in `pkgdown` website
> * [x]  in function documentations: note NULL, TRUE as code (by enclosing with ``), healthdata.gov as link, link to other functions by enclosing in ``
