### 2020-07-26 #358 @czeildi
@iecastro I reviewed your changes, looks great! I happily recommend approving this package.

I noticed one minor thing: you should no longer import `View` as you no longer use it: `#' @importFrom utils View` line should be deleted from `healthdatacsv-package.R`
