### 2020-07-27 #358 @karthik
Thanks very much for the review and acceptance recommendation @czeildi

The package looks to be in good shape to me as well. Since the second reviewer has gone missing, I'll try to get someone else to do a quick review and move this to acceptance.
