+++ R/healthdatacsv-package.R
@@ -12,6 +12,5 @@
 #' @importFrom tidyr unnest
 #' @importFrom tibble tibble
 #' @importFrom rlang .data
-#' @importFrom utils View
 #' @importFrom vroom vroom
 NULL
