+++ NAMESPACE
@@ -26,5 +26,4 @@ importFrom(stringr,regex)
 importFrom(stringr,str_detect)
 importFrom(tibble,tibble)
 importFrom(tidyr,unnest)
-importFrom(utils,View)
 importFrom(vroom,vroom)
