### 2020-08-24 #358 @iecastro
Thank you, @richfitz @karthik -  I appreciate everyone's time. I have reviewed the comments, and will start working on them soon.
