### 2020-10-24 5edc2b0
break up fetch_catalog
