+++ tests/testthat/test-fetch_catalog.R
@@ -1,5 +1,7 @@
 
+
 test_that("output is tibble", {
+  skip("skip")
   expect_is(fetch_catalog(), "tbl_df")
 })
 
