+++ tests/testthat/test-healthdata_api.R
@@ -1,13 +1,4 @@
 
-# api_call <- healthdata_api("data.json")
-
-#test_that("succesful status response code", {
-#  expect_equal(httr::status_code(api_call$response),
-#               200)
-#  }
-#)
-
-
 with_mock_api({
   api_call <- healthdata_api("data.json")
 
