+++ tests/testthat/test-healthdata_api.R
@@ -16,3 +16,5 @@ with_mock_api({
     expect_true(httpcache::hitCache("http://www.healthdata.gov/data.json"))
   })
 })
+
+
