+++ .travis.yml
@@ -9,7 +9,6 @@ r:
 
 cache: packages
 sudo: required
-dist: trusty
 
 warnings_are_errors: false
 
