+++ tests/testthat/test-fetch_catalog.R
@@ -1,6 +1,7 @@
 
 
 test_that("output is tibble", {
+  skip_on_cran()
   expect_is(fetch_catalog(), "tbl_df")
   expect_is(fetch_catalog(agency = "Centers for Disease Control and Prevention"),
             "tbl_df")
@@ -13,6 +14,7 @@ test_that("output is tibble", {
 
 
 test_that("message is displayed for wrong argument input", {
+  skip_on_cran()
   expect_message(fetch_catalog("CDC"))
   expect_message(fetch_catalog(keyword = "adpt"))
   expect_message(fetch_catalog(agency = "cdc", keyword = "adpt"))
@@ -24,6 +26,7 @@ test_that("message is displayed for wrong argument input", {
 
 
 test_that("keyword accepts regex pattern", {
+  skip_on_cran()
   expect_equal(
     nrow(fetch_catalog(keyword = "influenza|adopt")),
       nrow(fetch_catalog(keyword = "influenza")) +
