+++ .travis.yml
@@ -8,7 +8,11 @@ r:
   - devel
 
 cache: packages
-sudo: required
+
+addons:
+  apt:
+    packages:
+    - libcurl4-openssl-dev
 
 warnings_are_errors: false
 
