### 2020-11-05 #358 @iecastro
Hi @richfitz, @karthik  - Some comments below, addressing the review items.  I'm struggling with my travis build, and expanding the test suite.

@richfitz - if you have the time, would you mind attempting the review again?

Thanks, all!

> ### Review Comments
> Looked at 56bfd39 (Mon Jul 27 07:47:25 2020 -0400).
>
> I've not been able to complete the review ad the API is down, but the comments below should be addressed regardless. The three big issues are
>
> * The API is down - that's fine but the package should be able to cope with this gracefully.
> * The testing is incomplete.
> * The documentation is incomplete.
>
> Note that none of my comments really concern the _implementation_, which looks fine (modulo comments about dependencies, but that's a matter of taste).
>
> ## Problematic use of flakey API
> ```
> > devtools::test()
> Loading healthdatacsv
> Testing healthdatacsv
> ✔ |  OK F W S | Context
> ✖ |   0 1     | fetch_catalog [0.5 s]
> ────────────────────────────────────────────────────────────────────────────────
> test-fetch_catalog.R:3: error: output is tibble
> SSL certificate problem: certificate has expired
> Backtrace:
>   1. testthat::expect_is(fetch_catalog(), "tbl_df") tests/testthat/test-fetch_catalog.R:3:2
>   4. healthdatacsv::fetch_catalog()
>   5. healthdatacsv::healthdata_api("data.json") R/healthdata-api.R:31:2
>   6. httpcache::GET(data_url, user) R/healthdata-api.R:145:2
>   7. httr::GET(url, ...)
>   8. httr:::request_perform(req, hu$handle$handle)
>  10. httr:::request_fetch.write_memory(req$output, req$url, handle)
>  11. curl::curl_fetch_memory(url, handle = handle)
> ```
  * [x] __Is this the API being down, or an SSL Certificate issue?  I see the same when looking at my travis v. appveyor builds - not sure how to address this.  I think the error message is appropriate since it didn't get to make an API call; on the other hand, I haven't been able to find a way to mock a failed API call to test__

> All tests that use web requests should be opt-in, especially if the package will go to CRAN. Add `testthat::skip_on_cran()` to these test blocks

  * [x] __Completed__
>
> Similar, but harder, comments apply to the vignette if this is going on CRAN - I do not recommend at all trying to run a vignette that uses the internet or external service there. All alternative options are unpleasant though:

   * [x] __Completed.  I stripped the vignette to just a reference URL - will eventually replace with pkgdown article__
>
> * ask them not to run the vignette
> * precompile the vignette and included one that does not use the internet (numerous hacky ways of doing this)
> * include the vignette as a pkgdown article only (users then can't read it from R or offline)
> * mock out the api calls in the vignette (hard to maintain, ugly to write, you'll have to store the data in the package)
>
> ## Testing issues
> The testing is minimal to the point of view that it may not be worth having, for example in `test-fetch_catalogue.R`
>
> ```
> test_that("output is tibble", {
>   expect_is(fetch_catalog(), "tbl_df")
> })
> ```
>
> However, when looking at `fetch_catalogue()`, there's significant logic there that is never tested - just that we get a `tbl_df` out which is unavoidable if the function does not error. So this test could be replaced by checking that the function doesn't error (this is not the solution though!).
>
> Instead, split these functions into network-using and network free components. You're basically there already, and

  * [x] __I have split the function into network-using and network free components; and expanded the testing logic; although maybe not fully to the extend you had in mind?__

>
> ```r
> fetch_catalog <- function(agency = NULL,
>                           keyword = NULL) {
>
>   api_call <- healthdata_api("data.json")
>   fetch_catalogue_process(api_call, agency, keyword)
> }
> ```
>
> and focus the bulk of the testing on the second half. What happens if `agency` is not found in the data? What if the user provides a table rather than a string? Are multiple strings handled gracefully or is an error thrown etc.
>
> If you split the code at this point you don't even need to bother formally mocking the api call, you can just keep a "reasonable" data set in the package for testing with.
>
> Similar comments apply throughout.
>
> You may want to use mocking (I like the `mockery` package) for ensuring that your error messages in places like `healthdata_api` are constructed properly. One thing to watch for is that if `parsed` comes back without a `message` property, your message will be empty.
>
> The other coverage hole (`fetch_data`) look easy enough to fill but this is secondary to making sure that the tests that have been added are replaced with tests that validate more than the lack of errors.
>
> ### Documentation is minimal
> The documentation for `fetch_data`, the main entry point of the package I believe, reads - in full
  * [x] __I have expanded documentation__

> ```
> #' Fetch data from the healthdata.gov API
> #'
> #' This function will download data from
> #' [healthdata.gov](https://healthdata.gov/),
> #' if available in csv format
> ```
>
> This supposes that the user already knows what sort of data night be found there, and it's not clear what the restriction on csv is for. The `catalogue` is not explained relative to the `data`.
>
> (Other files have the same issues)
>
> In the `DESCRIPTION`, the `Description` field should probably be updated to reflect what the package is for, rather than how it works.
>
> ### Dependencies
> This seems to me an excessive number of dependencies to pull in for what is a nicely streamlined package. Things like `vroom`, `rlang`, `stringr` seem unwanted (is the speed difference of `vroom` at all noticable relative to the network cost of downloading the data). This is a purely personal view of mine - I'd not expect necessarily any changes here to accept this review. But if this was a package I, or someone on my team wrote, we'd be stripping down this list.
  * [x] __It's not just the speed difference that `vroom` provides, but also the indexing - it is specially helpful when dealing with datasets of ~500K rows__
  * [x] __The ability to leverage dependencies is what allows me to build something like this.  I didn't start with a package in mind; I was trying to find some data__
> On the other hand, `httpcache` looks like a nicely chosen dependency that streamlines the package. I've not been able to see it in action here though given the API is down.
>
> ### Minor comment
> Remove commented code (e.g., test-healthdata_api.R) - if part way through writing a test use `testthat::skip()`
   * [x] __Done__
> Please note I will be offline from 4-19 September. I can look again before then but will be unresponsive in those two weeks.
